fix: map screen ui
remove shadows from the stickies. make the map fill the screen while the

the map fill the space under the stickies while they appear

remove the logo from all screens except the first one. remove the text f

make the default background color #0e0e0e and change the light green to

remove the line above aislepilot on the list view screen and make the ba

make the back button higher and the aislepilot text centred in the list

remove the space above the heading aislepilot on the list screen

change the styling of the app to match walmart

remove the card icon in cart total